+++ Snowflake.Data/Client/SnowflakeDbConnection.cs
@@ -303,10 +303,10 @@ namespace Snowflake.Data.Client
                 // Otherwise when Dispose() is called, the close request would timeout.
                 _connectionState = ConnectionState.Closed;
                 logger.Error("Unable to connect: ", e);
+
                 if (e is SnowflakeDbException)
-                {
                     throw;
-                }
+
                 throw new SnowflakeDbException(
                         e,
                         SnowflakeDbException.CONNECTION_FAILURE_SSTATE,
@@ -318,9 +318,7 @@ namespace Snowflake.Data.Client
         internal void FillConnectionStringFromTomlConfigIfNotSet()
         {
             if (string.IsNullOrEmpty(ConnectionString))
-            {
                 ConnectionString = _tomlConnectionBuilder.GetConnectionStringFromToml();
-            }
         }
 
         public override Task OpenAsync(CancellationToken cancellationToken)
@@ -340,6 +338,7 @@ namespace Snowflake.Data.Client
                 OAuthClientSecret = OAuthClientSecret,
                 Token = Token
             };
+
             return SnowflakeDbConnectionPool
                 .GetSessionAsync(ConnectionString, sessionContext, cancellationToken)
                 .ContinueWith(previousTask =>
@@ -351,10 +350,10 @@ namespace Snowflake.Data.Client
                         _connectionState = ConnectionState.Closed;
                         logger.Error("Unable to connect", sfSessionEx);
                         throw new SnowflakeDbException(
-                           sfSessionEx,
-                           SnowflakeDbException.CONNECTION_FAILURE_SSTATE,
-                           SFError.INTERNAL_ERROR,
-                           "Unable to connect");
+                            sfSessionEx,
+                            SnowflakeDbException.CONNECTION_FAILURE_SSTATE,
+                            SFError.INTERNAL_ERROR,
+                            "Unable to connect");
                     }
                     else if (previousTask.IsCanceled)
                     {